Move pattern caller into utils
